moving deleteIdentity workflow to its own file
