+++ src/components/display/fewlines/ConfirmationBox/DeleteConfirmationBoxContent.tsx
@@ -1,13 +1,11 @@
-import Cookie from "js-cookie";
 import { useRouter } from "next/router";
 import React from "react";
 import styled from "styled-components";
 
 import { Button, ButtonVariant } from "../Button/Button";
 import { IdentityTypes } from "@lib/@types";
-import { HttpVerbs } from "@src/@types/core/HttpVerbs";
-import { fetchJson } from "@src/utils/fetchJson";
 import { getIdentityType } from "@src/utils/getIdentityType";
+import { deleteIdentity } from "@src/workflows/deleteIdentity";
 
 interface DeleteIdentityProps {
   setOpen: React.Dispatch<React.SetStateAction<boolean>>;
@@ -15,32 +13,6 @@ interface DeleteIdentityProps {
   type: IdentityTypes;
   value: string;
 }
-
-export const deleteIdentity = async (
-  userId: string,
-  type: IdentityTypes,
-  value: string,
-): Promise<void> => {
-  const requestData = {
-    userId,
-    type: getIdentityType(type),
-    value,
-  };
-
-  const deleteMessage = `${
-    getIdentityType(type) === IdentityTypes.EMAIL
-      ? "Email address"
-      : "Phone number"
-  } has been deleted`;
-
-  return fetchJson("/api/delete-identity", HttpVerbs.DELETE, requestData)
-    .then(() => {
-      Cookie.set("flashMessage", deleteMessage);
-    })
-    .catch((error) => {
-      throw error;
-    });
-};
 
 export const DeleteConfirmationBoxContent: React.FC<DeleteIdentityProps> = ({
   setOpen,
